+++ presentation/src/main/java/com/whiplash/presentation/SplashActivity.kt
@@ -7,6 +7,7 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.lifecycle.lifecycleScope
 import com.whiplash.presentation.databinding.ActivitySplashBinding
+import com.whiplash.presentation.login.LoginActivity
 import com.whiplash.presentation.util.ActivityUtils.navigateTo
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.delay
