+++ presentation/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/grey_900"
-    tools:context="com.whiplash.presentation.LoginActivity">
+    tools:context="com.whiplash.presentation.login.LoginActivity">
 
     <ImageView
         android:id="@+id/ivKakaoLogin"
